Add more modules (#193)
