+++ src/main.rs
@@ -1,215 +1,19 @@
-use ansi_term::Colour;
-use regex::Regex;
 use std::error::Error;
-use std::fs;
-use std::fs::File;
-use std::io::Write;
-use std::io::{self, BufRead};
-use std::path::Path;
-use std::process;
-use std::process::{Command, Stdio};
-use clap::ArgMatches;
 
+mod core;
+mod filesystem;
+mod fzf;
 mod option;
-
-fn read_lines<P>(filename: P) -> io::Result<io::Lines<io::BufReader<File>>>
-where
-    P: AsRef<Path>,
-{
-    let file = File::open(filename)?;
-    Ok(io::BufReader::new(file).lines())
-}
-
-fn gen_snippet(snippet: &str, line: &str) -> String {
-    if snippet.is_empty() {
-        line.to_string()
-    } else {
-        format!("{}{}", &snippet[..snippet.len() - 2], line)
-    }
-}
-
-fn limit_str(text: &str, length: usize) -> String {
-    if text.len() > length {
-        format!("{}…", &text[..length - 1])
-    } else {
-        format!("{:width$}", text, width = length)
-    }
-}
-
-fn parse_file(path: &str, stdin: &mut std::process::ChildStdin) {
-    let mut tags = String::from("");
-    let mut comment = String::from("");
-    let mut snippet = String::from("");
-
-    if let Ok(lines) = read_lines(path) {
-        for l in lines {
-            let line = l.unwrap();
-            if line.starts_with('%') {
-                tags = line;
-            } else if line.starts_with('#') {
-                comment = line;
-            } else if line.starts_with('$') {
-            }
-            // TODO
-            else if line.ends_with('\\') {
-                snippet = if !snippet.is_empty() {
-                    format!("{}{}", &snippet[..snippet.len() - 2], line)
-                } else {
-                    line
-                }
-            } else if line.is_empty() {
-            } else {
-                let full_snippet = gen_snippet(&snippet, &line);
-                match stdin.write(
-                    format!(
-                        "{col0}\t{col1}\t{col2}\t{tags}\t{comment}\t{snippet}\t\n",
-                        col0 = Colour::Red.paint(limit_str(&tags[2..], 16)),
-                        col1 = Colour::Blue.paint(limit_str(&comment[2..], 26)),
-                        col2 = &full_snippet,
-                        tags = &tags[2..],
-                        comment = &comment[2..],
-                        snippet = &full_snippet
-                    )
-                    .as_bytes(),
-                ) {
-                    Ok(_) => snippet = String::from(""),
-                    Err(_) => break,
-                }
-            }
-        }
-    }
-}
-
-fn extract_elements(argstr: &str) -> (&str, &str, &str) {
-    let mut parts = argstr.split('\t').skip(3);
-    let tags = parts.next().unwrap();
-    let comment = parts.next().unwrap();
-    let snippet = parts.next().unwrap();
-    (tags, comment, snippet)
-}
-
-fn call_fzf<F>(f: F) -> process::Output
-where
-    F: Fn(&mut process::ChildStdin) -> (),
-{
-    let mut child = Command::new("fzf")
-        .args(&[
-            "--height",
-            "100%",
-            "--preview-window",
-            "up:2",
-            "--with-nth",
-            "1,2,3",
-            "--delimiter",
-            "\t",
-            "--ansi",
-        ])
-        .args(&["--preview", "./navi preview {}"])
-        .stdin(Stdio::piped())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::inherit())
-        .spawn()
-        .unwrap();
-
-    let stdin = child
-        .stdin
-        .as_mut()
-        .ok_or("Child process stdin has not been captured!")
-        .unwrap();
-
-    f(stdin);
-
-    child.wait_with_output().unwrap()
-}
+mod parse;
+mod preview;
+mod shell;
 
 fn main() -> Result<(), Box<dyn Error>> {
-
     let matches = option::parse();
-    // println!("Value for config: {:#?}", matches);
 
     match matches.subcommand().0 {
-        "preview" => main_preview(&matches),
-        "widget" => main_shell(&matches),
-        _ => main_core(&matches),
-    }
-}
-
-fn main_shell(matches: &ArgMatches) -> Result<(), Box<dyn Error>> {
-    let file = match matches.subcommand().1.unwrap().value_of("shell").unwrap() {
-        "zsh" => "navi.plugin.zsh",
-        "fish" => "navi.plugin.fish",
-        _ => "navi.plugin.bash",
-    };
-
-    println!(
-        "{}/{}",
-        std::env::current_exe()
-            .unwrap()
-            .parent()
-            .unwrap()
-            .as_os_str()
-            .to_str()
-            .unwrap(),
-        file
-    );
-    Ok(())
-}
-
-fn main_preview(matches: &ArgMatches) -> Result<(), Box<dyn Error>> {
-    // println!("Value for config: {:#?}", matches.subcommand().1.unwrap().value_of("line").unwrap());
-
-    let (tags, comment, snippet) =
-        extract_elements(matches.subcommand().1.unwrap().value_of("line").unwrap()); // ("foo", "bar", "baz"); // extract_elements(&args[2]);
-    println!(
-        "{comment} {tags} \n{snippet}",
-        comment = Colour::Blue.paint(comment),
-        tags = Colour::Red.paint(format!("[{}]", tags)),
-        snippet = snippet
-    );
-
-    process::exit(0x0100)
-}
-
-fn main_core(_matches: &ArgMatches) -> Result<(), Box<dyn Error>> {
-    let output = call_fzf(|stdin| {
-        let paths = fs::read_dir("./cheats").unwrap();
-        for path in paths {
-            parse_file(
-                path.unwrap().path().into_os_string().to_str().unwrap(),
-                stdin,
-            );
-        }
-    });
-
-    if output.status.success() {
-        let raw_output = String::from_utf8(output.stdout)?;
-        let snippet = raw_output.split('\t').nth(5).unwrap();
-        let mut full_snippet = String::from(snippet);
-
-        let re = Regex::new(r"<(.*?)>").unwrap();
-        for cap in re.captures_iter(snippet) {
-            println!("{}", &cap[0]);
-            let bracketed_varname = &cap[0];
-            let varname = &bracketed_varname[1..bracketed_varname.len() - 1];
-
-            let output = call_fzf(|stdin| {
-                stdin.write_all(b"foo\n").unwrap();
-                stdin.write_all(b"bar\n").unwrap();
-                stdin.write_all(b"baz\n").unwrap();
-                stdin
-                    .write_all(format!("{}\n", varname).as_bytes())
-                    .unwrap();
-            });
-
-            let value = String::from_utf8(output.stdout).unwrap();
-            full_snippet = full_snippet.replace(bracketed_varname, &value[..value.len() - 1]);
-        }
-
-        Command::new("echo").arg(&full_snippet[..]).spawn()?;
-
-        Ok(())
-    } else {
-        let err = String::from_utf8(output.stderr)?;
-        panic!("External command failed:\n {}", err)
+        "preview" => preview::main(&matches),
+        "widget" => shell::main(&matches),
+        _ => core::main(&matches),
     }
 }
